requirements.txt
requests>=2.25.0
pydantic>=2.0.0
websocket-client>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0 